+++ classes/completion/custom_completion.php
@@ -34,7 +34,6 @@ use core_completion\activity_custom_completion;
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 class custom_completion extends activity_custom_completion {
-
     /**
      * Names of every custom rule this module can define.
      *