+++ lib.php
@@ -126,7 +126,7 @@ function modernvideoplayer_delete_instance($id): bool {
 
     $progressids = $DB->get_fieldset_select('modernvideoplayer_progress', 'id', 'modernvideoplayerid = ?', [$id]);
     if ($progressids) {
-        list($insql, $params) = $DB->get_in_or_equal($progressids, SQL_PARAMS_QM);
+        [$insql, $params] = $DB->get_in_or_equal($progressids, SQL_PARAMS_QM);
         $DB->delete_records_select('modernvideoplayer_segments', "progressid $insql", $params);
     }
     $DB->delete_records('modernvideoplayer_progress', ['modernvideoplayerid' => $id]);
@@ -160,6 +160,28 @@ function modernvideoplayer_save_files(stdClass $data, context_module $context): 
         0,
         ['subdirs' => 0, 'maxfiles' => 1]
     );
+
+    if (!empty($data->captions)) {
+        file_save_draft_area_files(
+            $data->captions,
+            $context->id,
+            'mod_modernvideoplayer',
+            'captions',
+            0,
+            ['subdirs' => 0, 'maxfiles' => 10]
+        );
+    }
+
+    if (!empty($data->chapters)) {
+        file_save_draft_area_files(
+            $data->chapters,
+            $context->id,
+            'mod_modernvideoplayer',
+            'chapters',
+            0,
+            ['subdirs' => 0, 'maxfiles' => 1]
+        );
+    }
 }
 
 /**
@@ -212,8 +234,11 @@ function modernvideoplayer_get_post_actions(): array {
 function modernvideoplayer_get_coursemodule_info($coursemodule): ?cached_cm_info {
     global $DB;
 
-    $instance = $DB->get_record('modernvideoplayer', ['id' => $coursemodule->instance],
-        'id, name, intro, introformat, requiredpercent, strictendvalidation');
+    $instance = $DB->get_record(
+        'modernvideoplayer',
+        ['id' => $coursemodule->instance],
+        'id, name, intro, introformat, requiredpercent, strictendvalidation'
+    );
     if (!$instance) {
         return null;
     }
@@ -283,7 +308,7 @@ function modernvideoplayer_pluginfile($course, $cm, $context, $filearea, $args, 
         return false;
     }
 
-    if (!in_array($filearea, ['video', 'poster'], true)) {
+    if (!in_array($filearea, ['video', 'poster', 'captions', 'chapters'], true)) {
         return false;
     }